Modifying dev and staging pipelines to workaround cloud build deficiency
Because Cloud Build does not trigger on refs/pull/ID/merge, the dev pipeline will build an image tagged with a commit SHA which is different from the image tag pushed by the staging pipeline. And this results in image-pull errors in k8s.

Fixed this by moving the image build and push steps to the staging pipeline.